Move deepestSharedRoot helper to separated file
