+++ src/bin/vuetype.ts
@@ -3,7 +3,7 @@
 import assert = require('assert')
 import path = require('path')
 import program = require('commander')
-import { globSync } from '../lib/file-util'
+import { globSync, deepestSharedRoot } from '../lib/file-util'
 import { findAndReadConfig } from '../lib/config'
 import { generate } from '../lib/generate'
 
@@ -22,23 +22,9 @@ if (program.args.length === 0) {
     return path.join(arg, '**/*.vue')
   })
 
-  const root = path.resolve(detectDeepestRoot(program.args))
+  const root = path.resolve(deepestSharedRoot(program.args))
   const config = findAndReadConfig(root)
   const options = config ? config.options : {}
   generate(globSync(patterns), options)
 }
 
-function detectDeepestRoot (pathNames: string[]): string {
-  assert(pathNames.length >= 1)
-
-  let root: string[] = pathNames[0].split(path.sep)
-  pathNames.slice(1).forEach(pathName => {
-    const dirs = pathName.split(path.sep)
-    dirs.forEach((dir, i) => {
-      if (root[i] !== dir) {
-        root = root.slice(0, i)
-      }
-    })
-  })
-  return root.join(path.sep)
-}
